Energy Monitoring Working

+++ Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib/appManager.h
@@ -10,7 +10,7 @@ typedef struct {
 
      connectionManager* conManager;
 //   energyMonitoringManager eManager;
-     unsigned long energy;
+     float energy;
      int switch_val;
      uint32_t waterLevel;
